ci: remove obsolete Expo Swift 6 workaround
Team-Machine-Session: p28251
