+++ .github/workflows/build-ios-kitchensink-app.yml
@@ -119,62 +119,6 @@ jobs:
         if: ${{ !steps.pre-check-has-build.outputs.cache-hit }}
         run: ${{ env.prebuild_command }}
 
-      # workaround for https://github.com/expo/expo/issues/42525
-      # remove once expo-modules-core ships a fix or expo SDK 56+ resolves it
-      - name: Patch Podfile for ExpoModulesCore Swift 6 and ContextMenuAuxiliaryPreview
-        if: ${{ !steps.pre-check-has-build.outputs.cache-hit }}
-        run: |
-          # workaround: expo-modules-core 55.x uses @MainActor in class inheritance
-          # clauses which requires Swift 6 mode. SWIFT_STRICT_CONCURRENCY=minimal
-          # suppresses concurrency warnings in swift 6 mode.
-          # workaround 2: ContextMenuAuxiliaryPreview uses deprecated transform: .default
-          # which is an error in Xcode 26 / Swift 6.2 - disable warnings-as-errors for it.
-          PODFILE="ios/Podfile"
-          if grep -q "ContextMenuAuxiliaryPreview" "$PODFILE" 2>/dev/null; then
-            echo "Podfile already patched (both ExpoModulesCore + ContextMenuAuxiliaryPreview), skipping"
-          elif [ -f "$PODFILE" ]; then
-            node -e "
-              const fs = require('fs');
-              const podfile = fs.readFileSync('$PODFILE', 'utf8');
-              const workaround = \`
-            # workaround: expo-modules-core 55.x requires Swift 6 mode with isolated
-            # conformances (SE-0470) for @MainActor in protocol conformance syntax.
-            installer.pods_project.targets.each do |target|
-              if target.name == 'ExpoModulesCore'
-                target.build_configurations.each do |build_config|
-                  build_config.build_settings['SWIFT_VERSION'] = '6'
-                  build_config.build_settings['SWIFT_STRICT_CONCURRENCY'] = 'minimal'
-                  flags = build_config.build_settings['OTHER_SWIFT_FLAGS'] || '$(inherited)'
-                  unless flags.include?('IsolatedConformances')
-                    build_config.build_settings['OTHER_SWIFT_FLAGS'] = "\#{flags} -enable-upcoming-feature IsolatedConformances"
-                  end
-                end
-              end
-              # workaround: ContextMenuAuxiliaryPreview uses deprecated transform: .default
-              # which becomes a build error in Xcode 26 / Swift 6.2 strict mode.
-              if target.name == 'ContextMenuAuxiliaryPreview'
-                target.build_configurations.each do |build_config|
-                  build_config.build_settings['SWIFT_TREAT_WARNINGS_AS_ERRORS'] = 'NO'
-                  build_config.build_settings['GCC_TREAT_WARNINGS_AS_ERRORS'] = 'NO'
-                end
-              end
-            end
-          \`;
-              const updated = podfile.replace(
-                /(post_install do \|installer\|.*?)(^\s+end\s*\nend)/ms,
-                '\$1' + workaround + '\$2'
-              );
-              if (updated === podfile) {
-                console.error('ERROR: could not find post_install block in Podfile');
-                process.exit(1);
-              }
-              fs.writeFileSync('$PODFILE', updated);
-              console.log('Patched Podfile with ExpoModulesCore Swift 6 and ContextMenuAuxiliaryPreview workarounds');
-            "
-          else
-            echo "WARNING: Podfile not found at $PODFILE"
-          fi
-
       - name: Cache Pods
         if: ${{ !steps.pre-check-has-build.outputs.cache-hit }}
         uses: actions/cache@v4
@@ -256,8 +200,6 @@ jobs:
             -sdk 'iphonesimulator' \
             -destination 'generic/platform=iOS Simulator' \
             -derivedDataPath build \
-            OTHER_SWIFT_FLAGS='$(inherited) -enable-upcoming-feature IsolatedConformances' \
-            SWIFT_STRICT_CONCURRENCY=minimal \
             RCT_REMOVE_LEGACY_ARCH=1 \
             | tee xcodebuild.log | xcpretty
