+++ code/kitchen-sink/app.json
@@ -37,7 +37,7 @@
     },
     "plugins": [
       "expo-font",
-      "./plugins/expo-modules-core-swift6",
+      "../packages/native-ci/plugins/context-menu-warnings",
       "./plugins/detox-metro-port",
       "./plugins/detox-skip-splash"
     ],
